% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_names.R
\name{cap_names}
\alias{cap_names}
\title{Name Capitalization for Frisbee Teams}
\usage{
cap_names(a, b)
}
\arguments{
\item{a}{Data Frame of the frisbee team}

\item{b}{Number of the column with given player names that wants to be
capitalized}
}
\value{
Names of the players in capitalized format
}
\description{
Converts the names within a given column to capitalize the first letter of
each player on a team
}
\examples{
cap_names(teamW, 10)
}
