% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catches.R
\name{catches}
\alias{catches}
\title{Number of catches by an individual player}
\usage{
catches(a, b)
}
\arguments{
\item{a}{Data frame of the frisbee team}

\item{b}{Name of the player}
}
\value{
Number of catches by that person
}
\description{
Number of catches by an individual player
}
\examples{
catches(teamW, "Tommy")
}
