% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catches-Graph.R
\name{catches_graph}
\alias{catches_graph}
\title{Using GGPlot2 functions to Graph Player Catches}
\usage{
catches_graph(data, ...)
}
\arguments{
\item{...}{Name of the players desired on the graph}
}
\value{
Plot of all listed players in the function
}
\description{
#' @param a Data frame of the frisbee team
}
\examples{
catches_g(teamW, "Riley", "Tommy", "Blake")
}
