% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drops.R
\name{drops}
\alias{drops}
\title{Number of drops by a player}
\usage{
drops(a, b)
}
\arguments{
\item{a}{Data frame of the frisbee team}

\item{b}{Name of the player}
}
\value{
Number of drops by that person
}
\description{
Number of drops by a player
}
\examples{
drops(teamW, "Ethan")
}
