% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goals.R
\name{goals}
\alias{goals}
\title{Number of goals scored by an individual}
\usage{
goals(a, b)
}
\arguments{
\item{a}{Data frame of the frisbee team}

\item{b}{Name of the person}
}
\value{
Number of goals scored by that person
}
\description{
Number of goals scored by an individual
}
\examples{
goals(teamW, "Riley")
}
