% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulls.R
\name{pulls}
\alias{pulls}
\title{Numbers of times for a Person Pulling the Frisbee}
\usage{
pulls(a, b)
}
\arguments{
\item{a}{A dataframe of the frisbee team}

\item{b}{Name of the puller}
}
\value{
Number of pulls
}
\description{
Take a player's name, and finds out the number of times they pulled
}
\examples{
pulls(teamW, "Tommy")
}
