% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Receiver-Stats.R
\name{receiver_stats}
\alias{receiver_stats}
\title{Overall stats for a given receiver on the team}
\usage{
receiver_stats(a, b)
}
\arguments{
\item{a}{Data frame of the frisbee team}

\item{b}{Name of the person}
}
\value{
Formatted summary of the player's drops, catches, drop rate, and

catch rate
}
\description{
Overall stats for a given receiver on the team
}
\examples{
receiver_stats(teamW, "Riley")
}
