% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teamW}
\alias{teamW}
\title{Data of Woodbury High School Ultimate Team}
\format{
A data frame with 6472 obs. of  20 variables
}
\source{
{https://www.ultianalytics.com/app/#/4729473996947456/players}
}
\usage{
teamW
}
\description{
Contains name of the tournament, opponent, time elapsed, passer/receiver,
and which players were playing on that line.
}
\examples{
data(teamW)
}
\keyword{datasets}
