% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thrower-Stats.R
\name{thrower_stats}
\alias{thrower_stats}
\title{Throwaway Stats}
\usage{
thrower_stats(a, b)
}
\arguments{
\item{b}{}
}
\value{
Formatted summary of the player's completion, throwaways, and their
rates
}
\description{
Throwaway Stats
}
\examples{
thrower_stats(teamW, "Ethan")
}
